% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpreSEd-package.R
\docType{package}
\name{ExpreSEd-package}
\alias{ExpreSEd}
\alias{ExpreSEd-package}
\title{ExpreSEd: ADS 8192 Differential Expression with DESeq2}
\description{
Provides a tool for differential gene analysis of lymph node regulatory T cells (Tregs) and conventional T cells (Tconv) isolated from four mouse tissues (fat, liver, lymph node, and skin). Tool includes functions that (1) filtering low expression gene counts, (2) preforms DESeq2 differential expression analysis, (3) applies log2fold-change shrinkage to the DESeq2 results, (4) summarizes the number of up/down/non-significant genes at given thresholds, (5) computes indepdenpt filtering diagonsitics, (6) visualizes results as a volcano plot, and (7) exportsthe results to TSV files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wilsonjewel27/ExpreSEd}
  \item Report bugs at \url{https://github.com/wilsonjewel27/ExpreSEd/issues}
}

}
\author{
\strong{Maintainer}: Wilson Jewel \email{jewel.wilson@stjude.org} (\href{https://orcid.org/0009-0009-6729-5476}{ORCID})

}
\keyword{internal}
